modules/backend/ai-ticket-processor/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# HTTP requests for Gemini API
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0 